@@ -20,7 +20,9 @@
     "@angular/platform-browser": "^2.4.0",
     "@angular/platform-browser-dynamic": "^2.4.0",
     "@angular/router": "^3.4.0",
+    "@types/node": "^7.0.5",
     "angular-in-memory-web-api": "^0.3.0",
+    "bootstrap": "^3.3.7",
     "core-js": "^2.4.1",
     "ng2-bootstrap": "^1.3.3",
     "rxjs": "^5.1.0",
@@ -28,6 +30,7 @@
   },
   "devDependencies": {
     "@angular/cli": "1.0.0-rc.1",
+    "@angular/compiler": "^2.4.8",
     "@angular/compiler-cli": "^2.4.0",
     "@types/jasmine": "2.5.38",
     "@types/node": "~6.0.60",
@@ -45,6 +48,6 @@
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.4.2",
-    "typescript": "~2.0.0"
+    "typescript": "^2.3.0-dev.20170228"
   }
 }
