@@ -1,8 +1,8 @@
-import { Component, OnInit }     from '@angular/core';
-import { Router }                from '@angular/router';
+import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 
-import { Observable }            from 'rxjs/Observable';
-import { Subject }               from 'rxjs/Subject';
+import { Observable } from 'rxjs/Observable';
+import { Subject } from 'rxjs/Subject';
 
 // Observable class extensions
 import 'rxjs/add/observable/of';
@@ -12,7 +12,7 @@ import 'rxjs/add/operator/catch';
 import 'rxjs/add/operator/debounceTime';
 import 'rxjs/add/operator/distinctUntilChanged';
 
-import { HeroSearchService }     from '../shared/hero-search.service';
+import { HeroSearchService } from '../shared/hero-search.service';
 import { Hero } from '../shared/hero.model';
 
 @Component({
@@ -58,7 +58,7 @@ export class HeroSearchComponent implements OnInit {
     }
 
     gotoDetail(hero: Hero): void {
-        let link = ['/detail', hero.id];
+        const link = ['/detail', hero.id];
         this.router.navigate(link);
     }
 
