@@ -1,11 +1,11 @@
-import { Injectable }       from '@angular/core';
-import { Http }             from '@angular/http';
+import { Injectable } from '@angular/core';
+import { Http } from '@angular/http';
 
-import                      'rxjs/add/operator/map';
-import { Observable }       from 'rxjs/Observable';
+import 'rxjs/add/operator/map';
+import { Observable } from 'rxjs/Observable';
 
 
-import { Hero }         from './hero.model';
+import { Hero } from './hero.model';
 
 @Injectable()
 export class HeroSearchService {
