@@ -9,6 +9,6 @@ describe('tour-of-heroes App', () => {
 
   it('should display message saying app works', () => {
     page.navigateTo();
-    expect(page.getParagraphText()).toEqual('app works!');
+    expect(page.getParagraphText()).toEqual('Tour of Heroes');
   });
 });
