@@ -1,4 +1,4 @@
-import { Component, Input, OnInit }        from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 
 import { Hero } from '../shared/hero.model';
 import { HeroService } from '../shared/hero.service';
